Restructured the hosts and variables
I wanted the user to be able to choose if this is a remote installation
or one being performed on localhost by picking a specific inventory
file. Part of this change was to separate the common variables into a
file of their own and use the inventory/groups_vars directory layout.