Fixed a bug with formatting the EFI boot partition
In cases where a machine is being set up more than once, even after
blowing away the partitions and re-creating, the filesystem module in
Ansible can decide the UEFI boot partition is already formatted. This
can leave files present from the prior install which causes pacstrap to
fail.

This fix enables the force parameter of filesystem for the boot
partition, avoiding this issue.